Reestructure files
