+++ winterfell_adapter/src/adapter/mod.rs
@@ -1,0 +1,2 @@
+pub mod public_inputs;
+pub mod air;
