+++ winterfell_adapter/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod adapter;
 pub mod field_element;
 pub mod examples;
+pub mod utils;
